refactor: hooks move to global
