+++ hooks/auth/useForm.ts
@@ -1,7 +1,9 @@
 import { useState, useEffect } from 'react';
-import { validatePassword } from '../utils/validation';
+import { validatePassword, validateUsername, validateEmail } from '@/utils/auth/validateForm';
 
 type Errors = {
+    emptyEmail: boolean;
+    emptyUsername: boolean;
     length: boolean;
     uppercase: boolean;
     lowercase: boolean;
@@ -13,6 +15,8 @@ type Errors = {
 // Customized Hook, used to manage form status and validation logic
 export const useForm = () => {
     // input value state
+    const [email, setEmail] = useState('');
+    const [firstName, setFirstName] = useState('');
     const [password, setPassword] = useState('');
     const [confirmPassword, setConfirmPassword] = useState('');
 
@@ -22,12 +26,16 @@ export const useForm = () => {
 
     // Whether the user has entered is used to control the display of error information.
     const [hasTyped, setHasTyped] = useState({
+        email: false,
+        firstName: false,
         password: false,
         confirmPassword: false,
     });
 
     // error state
     const [errors, setErrors] = useState<Errors>({
+        emptyEmail: false,
+        emptyUsername: false,
         length: false,
         uppercase: false,
         lowercase: false,
@@ -35,6 +43,22 @@ export const useForm = () => {
         specialChar: false,
         match: false,
     });
+
+    // Monitor user name changes for verification
+    useEffect(() => {
+        setErrors((prevErrors) => ({
+            ...prevErrors,
+            emptyEmail: !validateEmail(email),
+        }));
+    }, [email]);
+
+    // Monitor user name changes for verification
+    useEffect(() => {
+        setErrors((prevErrors) => ({
+            ...prevErrors,
+            emptyUsername: !validateUsername(firstName),
+        }));
+    }, [firstName]);
 
     // Monitor password and confirmation password changes for verification
     useEffect(() => {
@@ -48,6 +72,8 @@ export const useForm = () => {
 
     const isFormValid = () => {
         return (
+            !errors.emptyEmail &&
+            !errors.emptyUsername &&
             errors.length &&
             errors.uppercase &&
             errors.lowercase &&
@@ -62,10 +88,14 @@ export const useForm = () => {
 
     return {
         formValues: {
+            email,
+            firstName,
             password,
             confirmPassword,
         },
         setFormValues: {
+            setEmail,
+            setFirstName,
             setPassword,
             setConfirmPassword,
         },
